requirements.txt
airflow
snowflake-connector-python
pandas
requests
